49877: убрал тестовые изменения в экспортере

+++ src/service/app/Models/Exporter.cs
@@ -8,7 +8,6 @@ using System.IO;
 using System.Linq;
 using System.Text;
 using System.Threading;
-using System.Windows.Forms;
 using Castle.Components.Validator;
 using Common.Models;
 using Common.Models.Helpers;
@@ -1563,8 +1562,8 @@ where a.MailId in ({ids.Implode()})";
 
 				return;
 			}
-			//if (!userSettings.AllowDownloadUnconfirmedOrders)
-				//return;
+			if (!userSettings.AllowDownloadUnconfirmedOrders)
+				return;
 
 			session.CreateSQLQuery("delete from Logs.PendingOrderLogs where UserId = :userId")
 				.SetParameter("userId", user.Id)
@@ -1572,19 +1571,15 @@ where a.MailId in ({ids.Implode()})";
 
 			var addresses = Addresses.Select(a => (uint?)a.Id).ToArray();
 			var prices = session.Query<ActivePrice>().Select(p => p.Id).ToArray();
-			//MessageBox.Show(session.Query<Order>().ToArray().Length.ToString());
 			var orders = session.Query<Order>()
 				.Where(o => !o.Deleted
 					&& !o.Processed
 					&& !o.Submited
-					//&& o.UserId != user.Id
+					&& o.UserId != user.Id
 					&& addresses.Contains(o.AddressId))
 				.ToArray();
-			//MessageBox.Show(orders.Length.ToString());
 			orders = orders.Where(o => prices.Contains(new PriceKey(o.PriceList, o.RegionCode))).ToArray();
-			//MessageBox.Show(orders.Length.ToString());
 			var groups = orders.GroupBy(o => new { o.AddressId, o.PriceList, o.RegionCode });
-			//MessageBox.Show(groups.ToArray().Length.ToString());
 			foreach (var @group in groups) {
 				foreach (var order in group) {
 					session.Save(new PendingOrderLog(order, user, group.First().RowId));
